beginning version of menu
added -l and exit command

Added file runners and updated menu

updated input

added user input to menu

added docstrings

final version
